rest_api/pyproject.toml
[project]
name = "garden-rest-api"
version = "0.1.0"
description = "Rest API for garden project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django==4.1.13",
    "djongo==1.3.6",
    "redis==7.1.0",
    "cerberus==1.3.8",
    "pymongo==3.13.0",
    "pytz",
    "sqlparse==0.2.4",
    "paho-mqtt==1.6.1",
    "django-mqtt",
    "watchdog",
    "dnspython",
    "certifi",
    "djangorestframework>=3.15.1",
    "djangorestframework-simplejwt>=5.3.1",
    "pyjwt>=2.11.0",
    "requests>=2.32.5",
    "django-cors-headers>=4.5.0",
]

[tool.uv]
managed = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "ruff>=0.15.1",
]

[tool.ruff]
# Target version for Python
target-version = "py313"
line-length = 120

[tool.ruff.lint]
# Select rules to enable
select = [
"E",    # Errores de estilo (pycodestyle)
    "W",    # Advertencias de estilo (pycodestyle)
    "F",    # Errores lógicos (Pyflakes)
    "I",    # Orden de imports (isort)
    "N",    # Convenciones de nombres (pep8-naming)
    "B",    # Prevención de bugs comunes (flake8-bugbear)
    "UP",   # Actualización a sintaxis moderna (pyupgrade)
    "C4",   # Optimización de comprensiones (flake8-comprehensions)
    "PTH",  # Forzar uso de Pathlib (flake8-use-pathlib)
    "RUF",  # Reglas específicas de Ruff (muy útiles)
    "SIM",  # Simplificación de código (flake8-simplify)
    "ARG",  # Detectar argumentos no usados (flake8-unused-arguments)
]
ignore = [
    "E501",  # Line too long (handled by formatter)
    "N815",  # mixedCase variables in class scope (Django fields)
    "RUF012", # Mutable class attributes (standard Django patterns)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
